File Validation
It works, but it doesn't give as much diagnostic info to the user as I'd like.  Began separating it out into functions, so it can be called during the file reading phase (thereby having access to the file name during reporting).  ...I also think I should make an Error class at this point, but counter think that I'm overdesigning a simple Sudoku Solver.